Extract PM quality review action control
