+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -109,3 +109,4 @@
 | LWB-R102 | DPM wave action hook extraction | Hardened | `dpm-wave-command-center-panel.tsx` still owned rebalance action state, auto-loading proposed changes, campaign lifecycle/history/readiness/launch state, selected-campaign state, pending flags, and Gateway action calls inline beside display composition. That made Gateway-only campaign workflow behavior harder to test without mounting the full rebalance workspace and kept API imports in the panel layer. | Added `use-dpm-wave-command-center-actions.ts` to own Gateway-only rebalance preview/create/items/source-check/simulation/approval/stage/handoff/proof-pack actions plus campaign lifecycle, launch-history, preview-readiness, launch-package, and durable launch orchestration. The panel now composes the rebalance and campaign sections without importing DPM wave API helpers directly. Added focused hook coverage for selected-wave auto-load, bounded rebalance action call mapping, campaign evidence/readiness/launch call mapping, source-owned page size, and fail-closed no-launch behavior when Manage readiness is blocked. | Continue reducing DPM wave surfaces only where a slice separates Gateway orchestration from display composition or removes duplication. Keep campaign cohort discovery, membership, readiness calculation, idempotency construction, order generation, OMS, fills, settlement, execution, maker-checker approval, and global campaign discovery source-owned and Gateway-backed. | `src/features/workbench/use-dpm-wave-command-center-actions.ts`, `src/features/workbench/components/dpm-wave-command-center-panel.tsx`, `src/features/workbench/dpm-campaign-launch-history-constants.ts`, `tests/unit/use-dpm-wave-command-center-actions.test.ts`; validation via focused DPM wave tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R103 | Construction alternatives action hook extraction | Hardened | `construction-alternatives-panel.tsx` still owned Gateway generation, source-owned alternative selection, external OMS acknowledgement supportability loading, pending state, action messages, action errors, and selected-alternative enablement inline beside display composition. That kept API integration in the panel layer and made fail-closed selection behavior harder to test without rendering the full construction workspace. | Added `use-construction-alternatives-actions.ts` to own Gateway-only alternative generation, source-owned selection, external OMS acknowledgement supportability retrieval, pending/error/message state, derived construction model, portfolio id, and selected-alternative eligibility. The panel now composes summary, acknowledgement supportability, comparison, detail, and recommended-action cards without importing construction API helpers directly. Added hook coverage for Gateway acknowledgement loading, generation call mapping, selection call mapping, and fail-closed no-selection behavior before generation or when source readiness is blocked. | Continue reducing construction surfaces only where a slice separates Gateway orchestration from display composition or improves source-owned boundary tests. Keep optimizer logic, prices, target weights, hedge policy, eligible instruments, source readiness, selection truth, OMS acknowledgement, fills, settlement, execution status, and autonomous execution source-owned and Gateway-backed. | `src/features/workbench/use-construction-alternatives-actions.ts`, `src/features/workbench/components/construction-alternatives-panel.tsx`, `tests/unit/use-construction-alternatives-actions.test.ts`; validation via focused construction tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R104 | Report batch operations action hook extraction | Hardened | `report-batch-operations-panel.tsx` owned report-batch materialization, status refresh, bounded run-once, archived-document lookup, pending state, errors, batch identifiers, and terminal-status gating inline beside table and archive rendering. That kept Gateway-only reporting/archive orchestration coupled to presentation and made fail-closed behavior harder to test without rendering the full panel. | Added `use-report-batch-operations-actions.ts` to own Gateway-only report batch creation, source-owned status refresh, bounded worker pass, archive metadata lookup, pending/error state, derived batch id, terminal run gating, and archive reference state. The panel now composes actions, status, item table, and archive summary while delegating API orchestration to the hook. Added hook coverage for call mapping, status refresh, run-once refresh, archive-reference trimming, terminal gating, and no-call behavior when required identifiers are absent. | Continue reducing reporting surfaces only where it separates Gateway orchestration from display composition or removes duplicated report/archive state. Keep batch membership, readiness, idempotency, report generation state, archive truth, retention posture, execution, delivery, and communication workflows source-owned and Gateway-backed. | `src/features/workbench/use-report-batch-operations-actions.ts`, `src/features/workbench/components/report-batch-operations-panel.tsx`, `tests/unit/use-report-batch-operations-actions.test.ts`; validation via focused report batch tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
+| LWB-R105 | PM operating-quality review-action control extraction | Hardened | `pm-operating-quality-review-actions-card.tsx` mixed the bounded supervisory review-action form, preview/create readiness, create evidence rendering, ledger table, and detail evidence into one component after the review-action create slice. That made future control changes noisy and duplicated form-specific assertions inside the ledger/detail card test. | Added `pm-operating-quality-review-action-control.tsx` for the preview-before-create supervisory form, readiness strip, action buttons, and Manage create evidence. The review-actions card now delegates the command control and stays focused on ledger/detail rendering. Moved form behavior coverage into a focused control test while keeping a parent delegation test for composition. | Continue reducing PM operating-quality only where a slice separates command-control presentation from source-backed ledger/detail evidence or Gateway orchestration. Keep PM score calculation, fairness spread calculation, protected-class inference, PM ranking, HR/conduct decisions, client communication, approval, order, OMS, fills, settlement, and execution posture source-owned and Gateway-backed. | `src/features/workbench/components/pm-operating-quality-review-action-control.tsx`, `src/features/workbench/components/pm-operating-quality-review-actions-card.tsx`, `tests/unit/pm-operating-quality-review-action-control.test.tsx`, `tests/unit/pm-operating-quality-review-actions-card.test.tsx`; validation via focused PM-quality tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
